requirements.txt
Flask==2.3.2
flask-cors>=4.0.0
pandas>=1.5.0
requests>=2.31.0
anthropic>=0.18.0
werkzeug>=2.3.0
openpyxl>=3.0.0
reportlab>=3.6.0
google-genai>=1.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openai>=1.0.0